Feature/#755 - Open append() method on datanodes

+++ src/taipy/core/data/sql_table.py
@@ -116,25 +116,43 @@ class SQLTableDataNode(_AbstractSQLDataNode):
     def _get_base_read_query(self) -> str:
         return f"SELECT * FROM {self.properties[self.__TABLE_KEY]}"
 
+    def _do_append(self, data, engine, connection) -> None:
+        self.__insert_data(data, engine, connection)
+
     def _do_write(self, data, engine, connection) -> None:
+        self.__insert_data(data, engine, connection, delete_table=True)
+
+    def __insert_data(self, data, engine, connection, delete_table: bool = False) -> None:
+        """
+        Insert data into a SQL table.
+
+        Parameters:
+            data (List[Dict]): a list of dictionaries, where each dictionary represents a row of the table.
+            table: a SQLAlchemy object that represents a table.
+            connection: a SQLAlchemy connection to write the data.
+            delete_table (bool): indicates if the table should be deleted before inserting the data.
+        """
         table = self._create_table(engine)
         if isinstance(data, (modin_pd.DataFrame, pd.DataFrame)):
-            self._insert_dataframe(data, table, connection)
+            self.__insert_dataframe(data, table, connection, delete_table)
+            return
+
+        if isinstance(data, np.ndarray):
+            data = data.tolist()
+        if not isinstance(data, list):
+            data = [data]
+
+        if len(data) == 0:
+            self.__delete_all_rows(table, connection, delete_table)
+            return
+
+        if isinstance(data[0], (tuple, list)):
+            self.__insert_tuples(data, table, connection, delete_table)
+        elif isinstance(data[0], dict):
+            self.__insert_dicts(data, table, connection, delete_table)
+        # If data is a primitive type, it will be inserted as a tuple of one element.
         else:
-            if isinstance(data, np.ndarray):
-                data = data.tolist()
-            if not isinstance(data, list):
-                data = [data]
-            if len(data) == 0:
-                self._delete_all_rows(table, connection)
-            else:
-                if isinstance(data[0], (tuple, list)):
-                    self._insert_tuples(data, table, connection)
-                elif isinstance(data[0], dict):
-                    self._insert_dicts(data, table, connection)
-                # If data is a primitive type, it will be inserted as a tuple of one element.
-                else:
-                    self._insert_tuples([(x,) for x in data], table, connection)
+            self.__insert_tuples([(x,) for x in data], table, connection, delete_table)
 
     def _create_table(self, engine) -> Table:
         return Table(
@@ -143,37 +161,35 @@ class SQLTableDataNode(_AbstractSQLDataNode):
             autoload_with=engine,
         )
 
-    @staticmethod
-    def _insert_dicts(data: List[Dict], table: Any, connection: Any) -> None:
+    @classmethod
+    def __insert_dicts(cls, data: List[Dict], table: Any, connection: Any, delete_table: bool) -> None:
         """
         This method will insert the data contained in a list of dictionaries into a table. The query itself is handled
         by SQLAlchemy, so it's only needed to pass the correct data type.
         """
-        connection.execute(table.delete())
+        cls.__delete_all_rows(table, connection, delete_table)
         connection.execute(table.insert(), data)
 
-    @staticmethod
-    def _insert_dataframe(df: Union[modin_pd.DataFrame, pd.DataFrame], table: Any, connection: Any) -> None:
-        """
-        :param data: a pandas dataframe
-        :param table: a SQLAlchemy object that represents a table
-        :param connection: a SQLAlchemy connection to write the data
-        """
-        SQLTableDataNode._insert_dicts(df.to_dict(orient="records"), table, connection)
+    @classmethod
+    def __insert_dataframe(
+        cls, df: Union[modin_pd.DataFrame, pd.DataFrame], table: Any, connection: Any, delete_table: bool
+    ) -> None:
+        cls.__insert_dicts(df.to_dict(orient="records"), table, connection, delete_table)
 
-    @staticmethod
-    def _delete_all_rows(table, connection):
-        connection.execute(table.delete())
-
-    @staticmethod
-    def _insert_tuples(data: List[Union[Tuple, List]], table: Any, connection: Any) -> None:
+    @classmethod
+    def __insert_tuples(cls, data: List[Union[Tuple, List]], table: Any, connection: Any, delete_table: bool) -> None:
         """
         This method will look up the length of the first object of the list and build the insert through
         creation of a string of '?' equivalent to the length of the element. The '?' character is used as
         placeholder for a tuple of same size.
         """
-        connection.execute(table.delete())
+        cls.__delete_all_rows(table, connection, delete_table)
         markers = ",".join("?" * len(data[0]))
         ins = "INSERT INTO {tablename} VALUES ({markers})"
         ins = ins.format(tablename=table.name, markers=markers)
         connection.execute(ins, data)
+
+    @classmethod
+    def __delete_all_rows(cls, table: Any, connection: Any, delete_table: bool) -> None:
+        if delete_table:
+            connection.execute(table.delete())
